:recycle: refactor: homeService 제거 #70
